Rework address assignment logic
Add subnet4/6 options to the server configuration. If they are not
specified - assume point-to-point topology, otherwise - subnet.
This affects how address prefix is assigned at the client side. If the
subnet topology is used, then client gets an address with the subnet
prefix. Otherwise it is just /128 (or /32) with explicit route to reach
the server.

This allows wirebox suite to be used as both P2P tunneling software with
autoconfiguration and as a VPN.

Closes #13.